auth-sessions

+++ src/redux/tasks/operations.js
@@ -1,12 +1,18 @@
-import axios from 'axios';
 import { createAsyncThunk } from '@reduxjs/toolkit';
 
-axios.defaults.baseURL = 'https://task-pro-app-0x3n.onrender.com/api';
+import axios from '../../utils/backendAPI';
+import { checkRefreshAuthTokens } from '../../utils/backendAPI';
+
+// axios.defaults.baseURL = 'https://task-pro-app-0x3n.onrender.com/api';
 
 export const fetchBoards = createAsyncThunk(
   'boards/fetchBoards',
   async (_, thunkAPI) => {
     try {
+      // check and refresh tokens
+      const isValidTokens = await checkRefreshAuthTokens(thunkAPI);
+      if (!isValidTokens.status) throw isValidTokens.error;
+
       const { data } = await axios.get('/boards');
       return data.data;
     } catch (error) {
@@ -19,6 +25,10 @@ export const fetchOneBoard = createAsyncThunk(
   'boards/fetchOneBoard',
   async (id, thunkAPI) => {
     try {
+      // check and refresh tokens
+      const isValidTokens = await checkRefreshAuthTokens(thunkAPI);
+      if (!isValidTokens.status) throw isValidTokens.error;
+
       const { data } = await axios.get(`/boards/${id}`);
       return data.data;
     } catch (error) {
@@ -31,6 +41,10 @@ export const addBoard = createAsyncThunk(
   'boards/addBoard',
   async (newBoard, thunkAPI) => {
     try {
+      // check and refresh tokens
+      const isValidTokens = await checkRefreshAuthTokens(thunkAPI);
+      if (!isValidTokens.status) throw isValidTokens.error;
+
       const { data } = await axios.post('/boards', { ...newBoard });
       return data.data;
     } catch (error) {
@@ -43,6 +57,10 @@ export const editBoard = createAsyncThunk(
   'boards/editBoard',
   async ({ id, name, icon, background }, thunkAPI) => {
     try {
+      // check and refresh tokens
+      const isValidTokens = await checkRefreshAuthTokens(thunkAPI);
+      if (!isValidTokens.status) throw isValidTokens.error;
+
       const { data } = await axios.patch(`/boards/${id}`, {
         name,
         icon,
@@ -59,6 +77,10 @@ export const deleteBoard = createAsyncThunk(
   'boards/deleteBoard',
   async (id, thunkAPI) => {
     try {
+      // check and refresh tokens
+      const isValidTokens = await checkRefreshAuthTokens(thunkAPI);
+      if (!isValidTokens.status) throw isValidTokens.error;
+
       const { data } = await axios.delete(`/boards/${id}`);
       return data.data;
     } catch (error) {
@@ -71,6 +93,10 @@ export const createColumn = createAsyncThunk(
   'columns/createColumn',
   async (newColumn, thunkAPI) => {
     try {
+      // check and refresh tokens
+      const isValidTokens = await checkRefreshAuthTokens(thunkAPI);
+      if (!isValidTokens.status) throw isValidTokens.error;
+
       const { data } = await axios.post('/columns', { ...newColumn });
       return data.data;
     } catch (error) {
@@ -83,6 +109,10 @@ export const editColumn = createAsyncThunk(
   'columns/editColumn',
   async ({ id, name }, thunkAPI) => {
     try {
+      // check and refresh tokens
+      const isValidTokens = await checkRefreshAuthTokens(thunkAPI);
+      if (!isValidTokens.status) throw isValidTokens.error;
+
       const { data } = await axios.patch(`/columns/${id}`, { name });
       return data.data;
     } catch (error) {
@@ -95,6 +125,10 @@ export const deleteColumn = createAsyncThunk(
   'columns/deleteColumn',
   async (id, thunkAPI) => {
     try {
+      // check and refresh tokens
+      const isValidTokens = await checkRefreshAuthTokens(thunkAPI);
+      if (!isValidTokens.status) throw isValidTokens.error;
+
       const { data } = await axios.delete(`/columns/${id}`);
       return data.data;
     } catch (error) {
@@ -107,6 +141,10 @@ export const createTask = createAsyncThunk(
   'tasks/createTask',
   async (newTask, thunkAPI) => {
     try {
+      // check and refresh tokens
+      const isValidTokens = await checkRefreshAuthTokens(thunkAPI);
+      if (!isValidTokens.status) throw isValidTokens.error;
+
       const { data } = await axios.post('/tasks', { ...newTask });
       return data.data;
     } catch (error) {
@@ -119,6 +157,10 @@ export const editTask = createAsyncThunk(
   'tasks/editTask',
   async ({ id, name, description, priority, deadline }, thunkAPI) => {
     try {
+      // check and refresh tokens
+      const isValidTokens = await checkRefreshAuthTokens(thunkAPI);
+      if (!isValidTokens.status) throw isValidTokens.error;
+
       const { data } = await axios.patch(`/tasks/${id}`, {
         id,
         name,
@@ -135,15 +177,19 @@ export const editTask = createAsyncThunk(
 
 export const moveTask = createAsyncThunk(
   'tasks/moveTask',
-  async ({ taskId, fromColumnId, toColumnId }) => {
+  async ({ taskId, fromColumnId, toColumnId }, thunkAPI) => {
     try {
+      // check and refresh tokens
+      const isValidTokens = await checkRefreshAuthTokens(thunkAPI);
+      if (!isValidTokens.status) throw isValidTokens.error;
+
       const { data } = await axios.patch(`/tasks/${taskId}/move`, {
         fromColumnId,
         toColumnId,
       });
       return data.data;
     } catch (error) {
-      throw new Error(error.message);
+      return thunkAPI.rejectWithValue(error.message);
     }
   }
 );
@@ -152,6 +198,10 @@ export const deleteTask = createAsyncThunk(
   'tasks/deleteTask',
   async (id, thunkAPI) => {
     try {
+      // check and refresh tokens
+      const isValidTokens = await checkRefreshAuthTokens(thunkAPI);
+      if (!isValidTokens.status) throw isValidTokens.error;
+
       const { data } = await axios.delete(`/tasks/${id}`);
       return data.data;
     } catch (error) {
